Replace f64 Numbers With f32

+++ assembly/Config.ts
@@ -1,10 +1,10 @@
 export default class Config
 {
-  public static readonly width: f64 = 360;
-  public static readonly height: f64 = 240;
+  public static readonly width: u16 = 360;
+  public static readonly height: u16 = 240;
   public static readonly log: bool = false;
 
   public static readonly maxDepth: u8 = 50;
-  public static readonly pixelRatio: f64 = 1;
+  public static readonly pixelRatio: f32 = 1;
   // public static readonly samples: u16 = 500;
 }
